app/requirements.txt
aiodns==3.2.0
aioping==0.4.0
amqp==5.3.1
APScheduler==3.10.4
arrow==1.3.0
asgiref==3.7.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==23.2.0
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
celery==5.5.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Cython==3.0.10
Django==4.2.14
django-apscheduler==0.7.0
django-celery-beat==2.7.0
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-filter==23.5
django-js-asset==2.2.0
django-redis==5.4.0
django-rq==3.0.0
django-simple-history==3.7.0
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-spectacular==0.27.1
drf-writable-nested==0.7.0
et_xmlfile==2.0.0
executing==2.1.0
Flask==3.1.0
freezegun==1.5.1
gevent==24.2.1
greenlet==3.0.3
gTTS==2.5.1
gunicorn==23.0.0
h11==0.14.0
icecream==2.1.3
idna==3.6
imutils==0.5.4
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.5.2
MarkupSafe==3.0.2
numpy==1.26.3
opencv-python==4.9.0.80
openpyxl==3.1.5
packaging==24.0
pandas==2.2.0
pillow==10.3.0
prompt_toolkit==3.0.50
psutil==5.9.8
psycopg2==2.9.9
pycares==4.6.0
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pypng==0.20220715.0
pytesseract==0.3.10
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2023.4
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
redis==5.2.1
Redis-Sentinel-Url==1.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
rq==2.1.0
rq-dashboard==0.8.2.2
rq-scheduler==0.14.0
scipy==1.13.0
six==1.16.0
soupsieve==2.6
SpeechRecognition==3.10.3
sqlparse==0.4.4
tqdm==4.66.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.9.0
tzdata==2025.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.34.1
uWSGI==2.0.24
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
XlsxWriter==3.2.0
zope.event==5.0
zope.interface==6.4
